Delete old X86 stuff.
It seems that something has bit rotted, so this is still a work in progress.
